Enable NarrowCalculation error-prone check
This mostly matches IntelliJ's `integer multiplication implicitly cast
to long` inspection. While places changed do not look like bugs, so can
be viewed as false positives, this helps avoid IDE warnings and thus
makes spotting problems easier.

+++ core/trino-main/src/test/java/io/trino/spiller/TestBinaryFileSpiller.java
@@ -156,7 +156,7 @@ public class TestBinaryFileSpiller
         assertEquals(spillerStats.getTotalSpilledBytes() - spilledBytesBefore, spilledBytes);
         // At this point, the buffers should still be accounted for in the memory context, because
         // the spiller (FileSingleStreamSpiller) doesn't release its memory reservation until it's closed.
-        assertEquals(memoryContext.getBytes(), spills.length * FileSingleStreamSpiller.BUFFER_SIZE);
+        assertEquals(memoryContext.getBytes(), (long) spills.length * FileSingleStreamSpiller.BUFFER_SIZE);
 
         List<Iterator<Page>> actualSpills = spiller.getSpills();
         assertEquals(actualSpills.size(), spills.length);
